Relocated tests directory to resolve deployment issues
