+++ app/__tests__/index.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom'
-import Home from '../index'
+import Home from '../pages/index'
 
 describe('Test for pages/index.tsx', () => {
   beforeEach(() => {
